pyproject.toml
[project]
name = "rag-sandbox"
version = "0.1.0"
description = ""
authors = [
    {name = "Chandra Nandan",email = "chandranandan.chandrakar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<=3.12.8"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.18,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "langchain-redis (>=0.1.2,<0.2.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
